Resurrect `Microsoft.ML.PipelineInference`
This resurrects the `Microsoft.ML.PipelineInference` project. The
prpject was deleted as part of 37176a126f0016be5e68c13f67f8adadd594f40d.
The code included in this commit is a direct copy of the code as it was
in 71d58fa83f77abb630d815e5cf8aa9dd3390aa65, the commit right before it
was deleted.

This also re-adds the project to the Solution file.

NOTE: I could not test this, as current `master` doesnt build for me.